split shared types into domain files
