+++ packages/renderer-main/components/app-tab-strip.tsx
@@ -2,12 +2,8 @@ import { APP_TAB_STRIP_WIDE_WIDTH } from "@meru/shared/constants";
 import { ipc } from "@meru/shared/renderer/ipc";
 import { useConfig } from "@meru/shared/renderer/react-query";
 import type { AccountConfig } from "@meru/shared/schemas";
-import {
-  bookmarkableWorkspaceApps,
-  GMAIL_TAB_ID,
-  getTabStripWidth,
-  type TabState,
-} from "@meru/shared/types";
+import { GMAIL_TAB_ID, getTabStripWidth, type TabState } from "@meru/shared/tabs";
+import { bookmarkableWorkspaceApps } from "@meru/shared/workspace-apps";
 import { Button } from "@meru/ui/components/button";
 import {
   DropdownMenu,
