+++ packages/renderer-main/lib/stores.ts
@@ -2,7 +2,7 @@ import { ipc } from "@meru/shared/renderer/ipc";
 import { getConfig } from "@meru/shared/renderer/react-query";
 import { accountsSearchParam, trialDaysLeftSearchParam } from "@meru/shared/renderer/search-params";
 import type { AccountInstances } from "@meru/shared/schemas";
-import type { AccountTabsState } from "@meru/shared/types";
+import type { AccountTabsState } from "@meru/shared/tabs";
 import { toast } from "sonner";
 import { navigate } from "wouter/use-hash-location";
 import { create } from "zustand";
